tests: migrate ViewModels tests from MockK to Mokkery and move to commonTest
