+++ features/home/domain/src/commonTest/kotlin/dev/alvr/katana/features/home/domain/di/HomeDomainTestGraph.kt
@@ -1,10 +1,13 @@
 package dev.alvr.katana.features.home.domain.di
 
+import dev.alvr.katana.core.common.coroutines.KatanaDispatcher
 import dev.alvr.katana.core.tests.di.TestAppGraph
 import dev.alvr.katana.core.tests.di.TestAppScope
 import dev.alvr.katana.features.home.domain.repositories.HomeRepository
 import dev.alvr.katana.features.home.domain.usecases.HideWelcomeCardUseCase
+import dev.alvr.katana.features.home.domain.usecases.HideWelcomeCardUseCaseImpl
 import dev.alvr.katana.features.home.domain.usecases.ObserveWelcomeCardVisibilityUseCase
+import dev.alvr.katana.features.home.domain.usecases.ObserveWelcomeCardVisibilityUseCaseImpl
 import dev.zacsweers.metro.DependencyGraph
 import dev.zacsweers.metro.Provides
 import dev.zacsweers.metro.SingleIn
@@ -15,6 +18,18 @@ import dev.zacsweers.metro.createGraphFactory
 internal interface HomeDomainTestGraph : TestAppGraph {
     val hideWelcomeCardUseCase: HideWelcomeCardUseCase
     val observeWelcomeCardVisibilityUseCase: ObserveWelcomeCardVisibilityUseCase
+
+    @Provides
+    fun provideHideWelcomeCardUseCase(
+        dispatcher: KatanaDispatcher,
+        repository: HomeRepository,
+    ): HideWelcomeCardUseCase = HideWelcomeCardUseCaseImpl(dispatcher, repository)
+
+    @Provides
+    fun provideObserveWelcomeCardVisibilityUseCase(
+        dispatcher: KatanaDispatcher,
+        repository: HomeRepository,
+    ): ObserveWelcomeCardVisibilityUseCase = ObserveWelcomeCardVisibilityUseCaseImpl(dispatcher, repository)
 
     @DependencyGraph.Factory
     interface Factory {
