+++ features/home/ui/src/commonTest/kotlin/dev/alvr/katana/features/home/ui/di/ActivityUiTestGraph.kt
@@ -22,4 +22,4 @@ internal interface ActivityUiTestGraph : TestAppGraph {
     }
 }
 
-internal fun createActivityUiTestGraph() = createGraphFactory<ActivityUiTestGraph.Factory>().create().activityViewModel
+internal fun createActivityUiTestGraph() = createGraphFactory<ActivityUiTestGraph.Factory>().create()
